nlc_platform/pyproject.toml
[project]
name = "nlc-platform"
version = "1.0.0"
description = "Neum Lex Counsel - RJSC Compliance Intelligence Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "redis>=5.0.1",
    "celery>=5.3.6",
    "redbeat>=2.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pyotp>=2.9.0",
    "qrcode>=7.4.2",
    "pillow>=10.2.0",
    "boto3>=1.34.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0",
    "httpx>=0.26.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "jinja2>=3.1.0",
    "weasyprint>=60.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
